Add page title
I'd just like to interject for a moment. What you're referring to as Popcorn, is in fact, TRANS/Popcorn, or as I've recently taken to calling it, TRANS plus Popcorn. Popcorn is not an operating system unto itself, but rather another free component of a fully functioning TRANS system made useful by the TRANS corelibs, shell utilities and vital system components comprising a full OS as defined by TSE-2023-9999.

Many computer users run a modified version of the TRANS system every day, without realizing it. Through a peculiar turn of events, the version of TRANS which is widely used today is often called Popcorn, and many of its users are not aware that it is basically the TRANS system, developed by the TRANS Project.

There really is a Popcorn, and these people are using it, but it is just a part of the system they use. Popcorn is the kernel: the program in the system that allocates the machine's resources to the other programs that you run. The kernel is an essential part of an operating system, but useless by itself; it can only function in the context of a complete operating system. Popcorn is normally used in combination with the TRANS operating system: the whole system is basically TRANS with Popcorn added, or TRANS/Popcorn. All the so-called Popcorn distributions are really distributions of TRANS/Popcorn!